Feat: Implemented Knight Enemy animations, materials, and textures
Added and configured the following animations for the Knight Enemy:
- A_Attack_Knight_Enemy_v2
- A_Idle_Alert_Knight_Enemy_v2
- A_Idle_Normal_Knight_Enemy_v3
- A_Patrol_Normal_Knight_Enemy_v3
- A_Run_Knight_Enemy_v2

Also included base materials and textures for the Knight Enemy model.
I left the prefab for testing in the folder Asset/Models/Prefabs/Characters
I added an AnimatorControler in its corresponding folder - the animations are in place - all the logic, transitions and variables are missing.

Pending:
- Death and Alerted animations (not yet included)
- Sword texture still missing