feat (animation): Add new animations for Archer and Knight enemies
Adds a set of new animations to expand the behavior and state machines for the Archer and Knight enemy types.
Archer:
A_Archer_AlertedRun: A new run cycle for the alerted state
A_Archer_AlertedIdle_Prueba: New idle animation for the alerted state, to be used when the archer is aware of the player but not yet attacking.

Knight Enemy:
- A_IdleToPatrol_Knight_Enemy: A transition animation for when the knight moves from a stationary idle pose into its patrolling loop.